Changing types

+++ src/main/java/br/com/alelo/consumer/consumerpat/domain/base/EstablishmentCost.java
@@ -2,8 +2,10 @@ package br.com.alelo.consumer.consumerpat.domain.base;
 
 import br.com.alelo.consumer.consumerpat.entity.Consumer;
 
+import java.math.BigDecimal;
+
 public interface EstablishmentCost {
 
-    Consumer updateBalance(Double value, Integer cardNumber, Consumer consumer);
+    Consumer updateBalance(BigDecimal value, Integer cardNumber, Consumer consumer);
 
 }
